refactor compiles; first debug round

+++ st_spix/csrc/prop/update_prop_params.cu
@@ -14,7 +14,7 @@
 __host__ void update_prop_params(const float* img, const int* spix,
                                  superpixel_params* sp_params,
                                  superpixel_GPU_helper* sp_helper,
-                                 int* prev_means, int* prev_spix,
+                                 float* prev_means, int* prev_spix,
                                  const int npixels, const int nspix,
                                  const int nspix_buffer, const int nbatch,
                                  const int xdim, const int ydim, const int nftrs){
@@ -101,7 +101,7 @@ void sum_by_label(const float* img,
 __global__
 void calculate_mu_and_sigma(superpixel_params*  sp_params,
                             superpixel_GPU_helper* sp_helper,
-                            int* prev_means, int* prev_spix,
+                            float* prev_means, int* prev_spix,
                             const int nsuperpixel, const int nsuperpixel_buffer) {
 
     // -- update thread --
